# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API documentation for Vibecart project.",
        "title": "Vibecart API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "Check if API v1 is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "API v1 Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health.HealthResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if backend server is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Server Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health.HealthResponse"
                        }
                    }
                }
            }
        },
        "/test/aiven": {
            "get": {
                "description": "Execute SELECT 1 to verify database connection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Test Aiven MySQL Connection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health.DBTestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/health.DBTestErrorResponse"
                        }
                    }
                }
            }
        },
        "/test/cloudinary": {
            "get": {
                "description": "Upload small text file to verify Cloudinary credential",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Test Cloudinary Connection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health.CloudinaryTestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/health.CloudinaryTestErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "health.CloudinaryTestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid credentials"
                },
                "message": {
                    "type": "string",
                    "example": "Cloudinary connection failed"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "health.CloudinaryTestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Cloudinary connected successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "health.DBTestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "connection timeout"
                },
                "message": {
                    "type": "string",
                    "example": "Aiven connection failed"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "health.DBTestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Aiven connected successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "health.HealthResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Vibecart server is running"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    }
}